Refector: 자유게시판 전체 컴포넌트 코드 분리
